refactor: Reestructuración y optimización de los modelos de eventos municipales
Se ha realizado una refactorización significativa del código para mejorar la estructura y mantenibilidad del proyecto. Los principales cambios incluyen:

- Reorganización de directorios:
  - Creación de la carpeta 'evento' dentro de 'entidad_municipal_app/models' para alojar los modelos relacionados con eventos.
  - Migración de 'EventoMunicipal.py' y 'RegistroAsistencia.py' a la nueva estructura de carpetas.

- Actualización de modelos:
  - EventoMunicipal:
    - Implementación del manager personalizado 'EventoMunicipalManager' para manejar la creación de eventos con aforo.
    - Añadidas propiedades y métodos para gestionar cupos disponibles, inscripciones y notificaciones.
    - Mejoras en la gestión de estados del evento y validaciones adicionales.

  - RegistroAsistencia:
    - Creación del manager personalizado 'RegistroAsistenciaManager' para manejar consultas específicas.
    - Refactorización de métodos para gestionar inscripciones, estados y validaciones.
    - Implementación de métodos para actualizar estados y formatear fechas de inscripción.

- Adaptación de pruebas:
  - Actualización de los steps en 'control_asistencia.feature' para alinearse con la nueva estructura de modelos.
  - Incorporación de nuevos escenarios de prueba para cubrir los cambios realizados en los modelos.

- Mejoras generales:
  - Eliminación de importaciones circulares mediante referencias lazy en los modelos.
  - Optimización de consultas a la base de datos utilizando 'select_related' y transacciones atómicas para manejar concurrencia.
  - Implementación de manejo de excepciones personalizado con 'ErrorGestionEventos' para una mejor gestión de errores.

Estos cambios aseguran una mayor cohesión y separación de responsabilidades dentro del código, facilitando futuras expansiones y mantenimientos.
