+++ entidad_municipal_app/models/__init__.py
@@ -3,8 +3,8 @@ Modelos de la aplicación de entidad municipal.
 """
 
 from .EntidadMunicipal import EntidadMunicipal
-from .evento_municipal import EventoMunicipal
-from .registro_asistencia import RegistroAsistencia
+from .evento.evento_municipal import EventoMunicipal
+from .evento.registro_asistencia import RegistroAsistencia
 from .canales.canal_informativo import CanalInformativo, Suscripcion
 from .canales.noticia import Noticia
 from .canales.reaccion import Reaccion
