+++ entidad_municipal_app/models/evento/__init__.py
@@ -1,0 +1,8 @@
+"""
+Modelos relacionados con eventos municipales.
+"""
+
+from .evento_municipal import EventoMunicipal, ErrorGestionEventos
+from .registro_asistencia import RegistroAsistencia
+
+__all__ = ['EventoMunicipal', 'RegistroAsistencia', 'ErrorGestionEventos']
